+++ packages/ppx/test/css-support/content.t/run.t
@@ -83,6 +83,7 @@ If this test fail means that the module is not in sync with the ppx
   CSS.contentsRule([|`text({js|点|js})|], None);
   CSS.contentsRule([|`text({js|点|js})|], None);
   CSS.contentsRule([|`text({js|lola|js})|], None);
+  CSS.contentsRule([|`text({js|lola|js})|], None);
   CSS.contentsRule([|`text({js|''|js})|], None);
   CSS.contentsRule([|`text({js|' '|js})|], None);
   CSS.contentsRule([|`text({js|' '|js})|], None);